Split jsonrpc to internal and external

+++ core/src/storage/persistence/kv.rs
@@ -22,7 +22,7 @@ trait KvStorageBasic {
 #[allow(dead_code)]
 pub struct KvStorage {
     db: sled::Db,
-    cap: usize,
+    cap: u64,
     path: String,
 }
 
@@ -30,12 +30,12 @@ impl KvStorage {
     /// New KvStorage
     /// * cap: max_size in bytes
     /// * path: db file location
-    pub async fn new_with_cap_and_path<P>(cap: usize, path: P) -> Result<Self>
+    pub async fn new_with_cap_and_path<P>(cap: u64, path: P) -> Result<Self>
     where P: AsRef<std::path::Path> {
         let db = sled::Config::new()
             .path(path.as_ref())
             .mode(sled::Mode::HighThroughput)
-            .cache_capacity(cap as u64)
+            .cache_capacity(cap)
             .open()
             .map_err(Error::SledError)?;
         Ok(Self {
@@ -48,13 +48,13 @@ impl KvStorage {
     /// New KvStorage
     /// * cap: max_size in bytes
     /// * name: db file location
-    pub async fn new_with_cap_and_name(cap: usize, name: &str) -> Result<Self> {
+    pub async fn new_with_cap_and_name(cap: u64, name: &str) -> Result<Self> {
         Self::new_with_cap_and_path(cap, name).await
     }
 
     /// New KvStorage with default path
     /// default_path is `./`
-    pub async fn new_with_cap(cap: usize) -> Result<Self> {
+    pub async fn new_with_cap(cap: u64) -> Result<Self> {
         Self::new_with_cap_and_path(cap, "./data").await
     }
 
@@ -108,7 +108,7 @@ impl PersistenceStorageOperation for KvStorage {
         Ok(self.db.len() as u64)
     }
 
-    async fn max_size(&self) -> Result<usize> {
+    async fn max_size(&self) -> Result<u64> {
         Ok(self.cap)
     }
